+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 * Added function `newxml.parseXMLString()` for easy XML parsing.
 * Code readability improvements.
 * Basic Doctype handling added to the SAX parser.
+* Basic Doctype handling added to the DOM, with internal entity handling.
+
+### Known issues
+
+* Namespace URIs are not handled currently.
 
 # v0.2.1